Made progress on the Log Pearson Type III, 17B stats in the src/Stats.py file. Example 1from #17B is done... 2, 3, and 4 need to be completed still.
src/Storage.py uses Sympy to derive the formula to compute the exact volume of a storage structure, details in the module.

src/K_values.py is a hack at the moment because I haven't figured out the GammaInverse() function of the Sympy module yet. It's mostly just referencing the Table of K-Values that are available in Bulletin #17B (put into a csv of course).